src/requirements.txt
bleak==0.18.0
paho-mqtt==1.5.0
python-dotenv
bluepy
matplotlib
plotly
watchdog
pyjson5
pandas
jupyter
ipykernel
numpy
nbconvert